Move view logic to ViewHelper
